LIBFCREPO-1216: Updating tests and update command
moving majority of the test for updat to the stomp directory
edited the original file to remove the tests which use the messaging
also fixed the code styling in the update command and removed a static
method that's not being used anymore

https://umd-dit.atlassian.net/browse/LIBFCREPO-1216
